requirements-streamlit.txt
# Core ML packages
torch==2.0.1
torchvision==0.15.2
segmentation-models-pytorch==0.3.3

# Image processing
albumentations==1.3.1
opencv-python-headless==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
scikit-image>=0.16.1

# Visualization
matplotlib==3.7.2

# Streamlit and web components
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Utilities
tqdm==4.66.1